auth/requirements.txt
amqp==2.6.0
asgiref==3.2.7
beautifulsoup4==4.9.1
billiard==3.6.3.0
boto3==1.14.36
botocore==1.17.36
bs4==0.0.1
cachetools==4.1.1
celery==4.4.6
certifi==2020.4.5.1
cffi==1.14.1
chardet==3.0.4
Django==3.0.5
django-cors-headers==3.3.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
docutils==0.15.2
future==0.18.2
google-api-core==1.22.0
google-auth==1.20.0
google-cloud-core==1.4.1
google-cloud-storage==1.30.0
google-crc32c==0.1.0
google-resumable-media==0.7.0
googleapis-common-protos==1.52.0
idna==2.9
jmespath==0.10.0
kombu==4.6.11
loguru==0.5.3
lxml==4.5.1
Pillow==7.1.2
protobuf==3.12.4
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
python-http-client==3.2.7
pytz==2020.1
redis==3.5.3
requests==2.23.0
rsa==4.6
s3transfer==0.3.3
sendgrid==6.3.1
sentry-sdk==0.14.4
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
urllib3==1.25.9
vine==1.3.0
